Align Stainless with wrapping Int semantics; disable termination checks
- Disable check-measures and strict-arithmetic in stainless.conf, which
  the plugin honors (confirmed with positive and negative controls for
  each): neither first-class refinement types nor LiquidTyper prove
  termination or absence of overflow, so this aligns the measured
  verification work and the integer semantics across platforms.
- Convert postuple, rational, vec and mergeSort (and their bases) from
  BigInt to plain Int, and drop sumnat's wrapping calls: all three
  platforms now reason over the same wrapping machine integers, without
  BigInt(...) or wrapping(...) token noise.
- mergeSort: remove the decreases measures and restore the first-class
  sum-only splitAt contract, both of which existed only for termination
  checking.
- sqrt: use the real scala.math.sqrt body under @extern; the require is
  still enforced at call sites (verified with a negative test), matching
  the first-class and schmid variants. Plain scala.math.sqrt does not
  extract ("missing dependencies"); bolts only uses it in unverified
  code.

+++ evaluation/sources/mergeSort-stainless-base.scala
@@ -2,20 +2,20 @@ import stainless.lang._
 
 object MergeSort {
 
-  def safeDiv(x: BigInt, y: BigInt): BigInt = {
-    (??? : BigInt)
+  def safeDiv(x: Int, y: Int): Int = {
+    (??? : Int)
   }
 
   trait SafeSeq[T] {
-    def len: BigInt = {
-      (??? : BigInt)
+    def len: Int = {
+      (??? : Int)
     }
 
-    def apply(i: BigInt): T = {
+    def apply(i: Int): T = {
       (??? : T)
     }
 
-    def splitAt(i: BigInt): (SafeSeq[T], SafeSeq[T]) = {
+    def splitAt(i: Int): (SafeSeq[T], SafeSeq[T]) = {
       (??? : (SafeSeq[T], SafeSeq[T]))
     }
 
@@ -27,7 +27,7 @@ object MergeSort {
       (??? : T)
     }
 
-    def take(n: BigInt): SafeSeq[T] = {
+    def take(n: Int): SafeSeq[T] = {
       (??? : SafeSeq[T])
     }
